pyproject.toml
[project]
name = "dho-analysis"
version = "0.1.0"
description = ""
authors = [
    {name = "Björn",email = "bjoernwe@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "polars (>=1.22.0,<2.0.0)",
    "sentence-transformers (>=3.4.1,<4.0.0)",
    "sklearn-sfa (>=0.1.6,<0.2.0)",
    "matplotlib (>=3.10.0,<4.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "joblib (>=1.4.2,<2.0.0)",
]

[tool.poetry]
#packages = [{include = "dho_analysis", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
